requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
SQLAlchemy-Serializer
